Lint

+++ dispersing/kaitai_utilities.py
@@ -1,5 +1,7 @@
 import numpy as np
+
 from .fast_utilities import unpack_sprite_algo3
+
 
 def common_attributes(obj_collection):
     shared_attributes = set()
@@ -7,24 +9,27 @@ def common_attributes(obj_collection):
         obj_attributes = set()
         for aname in (_ for _ in dir(obj) if not _.startswith("_")):
             attr = getattr(obj, aname)
-            if callable(attr): continue
-            obj_attributes.add( (aname, type(attr)) )
+            if callable(attr):
+                continue
+            obj_attributes.add((aname, type(attr)))
         if len(shared_attributes) == 0:
             shared_attributes.update(obj_attributes)
         else:
             shared_attributes.intersection_update(obj_attributes)
     return [_[0] for _ in sorted(shared_attributes)]
 
+
 def collect_attributes(obj_collection, attr_list):
-    values = {_:[] for _ in attr_list}
+    values = {_: [] for _ in attr_list}
     for obj in obj_collection:
         for attr in attr_list:
             values[attr].append(getattr(obj, attr))
     return values
 
+
 def py_unpack_sprite_algo3(s):
     s = np.frombuffer(s, dtype="u1")
-    history = np.zeros(0x1000, dtype='u1')
+    history = np.zeros(0x1000, dtype="u1")
     history[:] = 0xFE
     it = 0xFEE
     si = 0
@@ -41,7 +46,7 @@ def py_unpack_sprite_algo3(s):
             pos += 1
             d.append(v)
             history[it] = v
-            it = (it+1) & 0xFFF
+            it = (it + 1) & 0xFFF
         else:
             v1 = s[pos]
             pos += 1
@@ -55,6 +60,6 @@ def py_unpack_sprite_algo3(s):
                 v = history[bx]
                 d.append(v)
                 history[it] = v
-                it = (it+1) & 0xFFF
+                it = (it + 1) & 0xFFF
                 cx += 1
-    return np.array(d, dtype='uint8')
+    return np.array(d, dtype="uint8")
